Tasks 3 and 4 done
Tasks 3 and 4 are done.

LIMITATIONS:
For some reason there are bugs wiht the camera intent. Even if the pic is taken it restarts the intent and I don't know why. Culd not find anything on the internet as well. Sometimes it works and if it doesn't you just have to press back and it will be there